requirements.txt
pip~=22.0.4
attrs~=19.3.0
distro~=1.6.0
toml~=0.10.2
cryptography~=36.0.0
py~=1.11.0
lxml~=4.7.1
Pillow~=9.1.0
docutils~=0.18.1
tornado~=6.1
Jinja2~=3.0.3
pytz~=2019.3
idna~=2.8
multidict~=6.0.2
Babel~=2.9.1
aiohttp~=3.8.1
certifi~=2019.9.11
aiogram~=2.20
yarl~=1.7.2
aiosignal~=1.2.0
frozenlist~=1.3.0
Brotli~=1.0.9
pytest~=5.3.5
six~=1.16.0
urllib3~=1.25.6
chardet~=3.0.4
setuptools~=59.6.0
requests~=2.26.0
defusedxml~=0.7.1
cffi~=1.15.0
PyQt5~=5.15.6
future~=0.18.2
olefile~=0.46
beautifulsoup4~=4.11.0
html5lib~=1.1
MarkupSafe~=2.0.0
pluggy~=0.13.1
packaging~=20.1
pexpect~=4.8.0
argcomplete~=2.0.0
pycurl~=7.45.1
wcwidth~=0.1.8
Pygments~=2.10.0
pyparsing~=2.4.6
ply~=3.11
pycparser~=2.20
zipp~=2.2.0
python-dotenv~=0.19.0